+++ R/svyjskm.R
@@ -14,6 +14,7 @@
 #' @param pval.size numeric value specifying the p-value text size. Default is 4.
 #' @param pval.coord numeric vector, of length 2, specifying the x and y coordinates of the p-value. Default values are NULL
 #' @param pval.testname logical: add '(Log-rank)' text to p-value. Default = F
+#' @param marks logical: should censoring marks be added?
 #' @param hr logical: add the Hazard Ratio to the plot?, Default: FALSE
 #' @param hr.size numeric value specifying the Hazard Ratio text size. Default is 2.
 #' @param hr.coord numeric vector, of length 2, specifying the x and y coordinates of the Hazard Ratio. Default values are NULL
@@ -38,6 +39,7 @@
 #' @param nejm.infigure.ratioh Ratio of infigure height to total height, Default = 0.5
 #' @param nejm.infigure.ylim y-axis limit of infigure, Default = c(0,1)
 #' @param surv.by breaks unit in y-axis, default = NULL(ggplot default)
+#' @param nejm.surv.by breaks unit in y-axis in nejm figure, default = NULL(ggplot default)
 #' @param ... PARAM_DESCRIPTION
 #' @return plot
 #' @details DETAILS
@@ -74,6 +76,7 @@ svyjskm <- function(sfit,
                     pval.size = 4,
                     pval.coord = c(NULL, NULL),
                     pval.testname = F,
+                    marks = FALSE, 
                     hr = FALSE,   # 수정 중 
                     hr.size = 2,  # 수정 중
                     hr.coord = c(NULL, NULL),  # 수정 중_
@@ -97,8 +100,9 @@ svyjskm <- function(sfit,
                     nejm.infigure.ratioh = 0.5,
                     nejm.infigure.ylim = c(0, 1),
                     surv.by = NULL,
+                    nejm.surv.by = NULL,
                     ...) {
-  surv <- strata <- lower <- upper <- NULL
+  n.censor <- surv <- strata <- lower <- upper <- NULL
 
   if (!is.null(theme) && theme == "nejm") legendposition <- legendposition
   if (is.null(timeby)) {
@@ -118,7 +122,53 @@ svyjskm <- function(sfit,
       ci <- "varlog" %in% names(sfit)
     }
   }
+  
+  
+  if (inherits(sfit, "svykmlist")) {
+    temp <- "varlog" %in% names(sfit[[1]])
+    if (temp & marks){
+      warning("Cannot mark when se=T")
+      marks <- F
+    }
+  } else if (inherits(sfit, "svykm")) {
+    temp <- "varlog" %in% names(sfit)
+    if (temp & marks){
+      warning("Cannot mark when se=T")
+      marks <- F
+    }
+  }
+  
 
+  
+  ##여기서 n.censor 시작할꺼야
+  #if (marks == T){
+    if (is.null(design)){
+      design <- tryCatch(get(as.character(attr(sfit, "call")$design)), error = function(e) e)
+      if ("error" %in% class(design)) {
+        warning("Design not found, can't mark it")
+        marks <- F
+      }
+    }
+  if(!is.null(design)){
+    var.time <- as.character(formula(sfit)[[2]][[2]])
+    var.event <- as.character(formula(sfit)[[2]][[3]])
+    if (length(var.event) > 1) {
+      var.event <- setdiff(var.event, as.character(as.symbol(var.event)))
+      var.event <- var.event[sapply(var.event, function(x) {
+        "warning" %in% class(tryCatch(as.numeric(x), warning = function(w) w))
+      })]
+    }
+    design1 <- design
+    form <- formula(sfit)
+    
+    fit2 <- survfit(form, data = design1$variables)
+  }
+    
+    
+  #}
+  
+  
+  
 
   if (ci & !is.null(cut.landmark)) {
     if (is.null(design)) {
@@ -200,10 +250,21 @@ svyjskm <- function(sfit,
             df2$med <- df2[closest_index, "time"]
             return(df2)
           }))
+          # mat <- cbind(fit2$time, fit2$n.censor)
+          # colnames(mat) <- c("time", "n.censor")
+          # mat_df <- as.data.frame(mat)
+          # df <- merge(x = df, y = mat_df, by = "time", all.x = TRUE)
         } else {
           df <- do.call(rbind, lapply(names(sfit), function(x) {
             data.frame("strata" = x, "time" = sfit[[x]]$time, "surv" = sfit[[x]]$surv)
           }))
+         # if (marks){
+            # mat <- cbind(fit2$time, fit2$n.censor)
+            # colnames(mat) <- c("time", "n.censor")
+            # mat_df <- as.data.frame(mat)
+            # df <- merge(x = df, y = mat_df, by = "time", all.x = TRUE)
+            #print(df2)
+          #}
         }
       }
     }
@@ -282,6 +343,17 @@ svyjskm <- function(sfit,
     }
   }
 
+  
+  if (!is.null(design)){
+    mat <- cbind(fit2$time, fit2$n.censor)
+    colnames(mat) <- c("time", "n.censor")
+    mat_df <- as.data.frame(mat)
+    df <- merge(x = df, y = mat_df, by = "time", all.x = TRUE)
+  }else{
+    df$n.censor <- NA
+  }
+  
+    
   m <- max(nchar(ystratalabs))
 
 
@@ -301,9 +373,9 @@ svyjskm <- function(sfit,
   # Final changes to data for survival plot
   levels(df$strata) <- ystratalabs
   zeros <- if (med == T & is.null(cut.landmark)) {
-    data.frame("strata" = factor(ystratalabs, levels = levels(df$strata)), "time" = 0, "surv" = 1, "med" = 0.5)
+    data.frame("strata" = factor(ystratalabs, levels = levels(df$strata)), "time" = 0, "surv" = 1, "med" = 0.5, "n.censor"=NA)
   } else {
-    data.frame("strata" = factor(ystratalabs, levels = levels(df$strata)), "time" = 0, "surv" = 1)
+    data.frame("strata" = factor(ystratalabs, levels = levels(df$strata)), "time" = 0, "surv" = 1, "n.censor"=NA)
   }
 
   if (ci) {
@@ -337,7 +409,7 @@ svyjskm <- function(sfit,
   } else {
     linetype <- c("solid", "solid", "solid", "solid", "solid", "solid", "solid", "solid", "solid", "solid", "solid")
   }
-
+  
   # Scale transformation
   # ::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::
   surv.scale <- match.arg(surv.scale)
@@ -408,7 +480,12 @@ svyjskm <- function(sfit,
       scale_linetype_manual(name = ystrataname, values = linetype, labels = ystratalabs)
   }
 
-
+  if (marks == TRUE) {
+    p <- p + geom_point(data = subset(df, n.censor >= 1), aes(x = time, y = surv, colour = strata), shape = 3)
+  }
+  
+  
+  
   # Add median value
   if (med == TRUE & is.null(cut.landmark)) {
     df3 <- df[-c(1, 2), ]
@@ -839,10 +916,19 @@ svyjskm <- function(sfit,
   # Plotting the graphs #
   #######################
   if (!is.null(theme) && theme == "nejm") {
-    p2 <- p1 + coord_cartesian(ylim = nejm.infigure.ylim) + theme(
-      axis.title.x = element_blank(), axis.title.y = element_blank(),
-      axis.text = element_text(size = 10 * nejm.infigure.ratiow) + scale_y_continuous(limits = nejm.infigure.ylim, breaks = waiver(), labels = scale_labels)
-    ) + guides(colour = "none", linetype = "none")
+    if (!is.null(nejm.surv.by)) {
+      p2 <- p1 + coord_cartesian(ylim = nejm.infigure.ylim) + theme(
+        axis.title.x = element_blank(), axis.title.y = element_blank(),
+        axis.text = element_text(size = 10 * nejm.infigure.ratiow) 
+      ) + guides(colour = "none", linetype = "none")+ scale_y_continuous(limits = nejm.infigure.ylim, breaks = seq(nejm.infigure.ylim[1], nejm.infigure.ylim[2], by = nejm.surv.by), labels = scale_labels)
+      } else {
+      p2 <- p1 + coord_cartesian(ylim = nejm.infigure.ylim) + theme(
+        axis.title.x = element_blank(), axis.title.y = element_blank(),
+        axis.text = element_text(size = 10 * nejm.infigure.ratiow) + scale_y_continuous(limits = nejm.infigure.ylim, breaks = waiver(), labels = scale_labels)
+      ) + guides(colour = "none", linetype = "none")
+    }
+    
+    
     p <- p + patchwork::inset_element(p2, 1 - nejm.infigure.ratiow, 1 - nejm.infigure.ratioh, 1, 1, align_to = "panel")
   }